- Minor fixes and added default 4" launch image to allow for larger screen operation.

+++ RunShoeMileage/UserDistanceSetting.h
@@ -36,7 +36,7 @@
 
 + (int) getSelectedShoe;
 
-+ (void) setSelectedShoe:(int)shoeIndex;
++ (void) setSelectedShoe:(NSInteger)shoeIndex;
 
 
 @end
